+++ cli/Cargo.toml
@@ -43,9 +43,6 @@ trust-dns-resolver = { workspace = true }
 multihash-codetable = { workspace = true, features = ["digest", "serde", "sha2"] }
 jsonrpsee = { workspace = true, features = ["async-client", "client-ws-transport-native-tls"] }
 prost = "0.12"
-schnorrkel = "0.11.4"
-blake2 = "0.10.4"
-bs58 = { version = "0.5.0", features = ["alloc"] }
 ss58-registry = { version = "1.34.0", default-features = false }
 
 [build-dependencies]
