+++ cli/src/commands/authorities.rs
@@ -22,6 +22,8 @@ use std::collections::{HashMap, HashSet};
 use subp2p_explorer::{
     peer_behavior::PeerInfoEvent,
     transport::{TransportBuilder, MIB},
+    util::crypto::sr25519,
+    util::ss58::{ss58hash, to_ss58},
     Behaviour, BehaviourEvent,
 };
 
@@ -32,31 +34,6 @@ pub async fn client(url: Url) -> Result<Client, Box<dyn std::error::Error>> {
     Ok(Client::builder()
         .max_buffer_capacity_per_subscription(4096)
         .build_with_tokio(sender, receiver))
-}
-
-/// Polkadot crypto module.
-mod sr25519 {
-    /// Public key for sr25519 keypair implementation.
-    pub type PublicKey = [u8; 32];
-
-    /// Signature generated by signing.
-    pub type Signature = [u8; 64];
-
-    /// Verify that the signature of the message was generated with the public key.
-    pub fn verify<M: AsRef<[u8]>>(sig: &Signature, message: M, pubkey: &PublicKey) -> bool {
-        const SIGNING_CTX: &[u8] = b"substrate";
-
-        let Ok(signature) = schnorrkel::Signature::from_bytes(sig) else {
-            return false;
-        };
-        let Ok(public) = schnorrkel::PublicKey::from_bytes(pubkey) else {
-            return false;
-        };
-
-        public
-            .verify_simple(SIGNING_CTX, message.as_ref(), &signature)
-            .is_ok()
-    }
 }
 
 /// Call the runtime API of the target node to retrive the current set
@@ -574,37 +551,6 @@ impl PeerInfo {
             }
         }
     }
-}
-
-fn ss58hash(data: &[u8]) -> Vec<u8> {
-    use blake2::{Blake2b512, Digest};
-    const PREFIX: &[u8] = b"SS58PRE";
-
-    let mut ctx = Blake2b512::new();
-    ctx.update(PREFIX);
-    ctx.update(data);
-    ctx.finalize().to_vec()
-}
-
-fn to_ss58(key: &sr25519::PublicKey, version: u16) -> String {
-    // We mask out the upper two bits of the ident - SS58 Prefix currently only supports 14-bits
-    let ident: u16 = version & 0b0011_1111_1111_1111;
-    let mut v = match ident {
-        0..=63 => vec![ident as u8],
-        64..=16_383 => {
-            // upper six bits of the lower byte(!)
-            let first = ((ident & 0b0000_0000_1111_1100) as u8) >> 2;
-            // lower two bits of the lower byte in the high pos,
-            // lower bits of the upper byte in the low pos
-            let second = ((ident >> 8) as u8) | ((ident & 0b0000_0000_0000_0011) as u8) << 6;
-            vec![first | 0b01000000, second]
-        }
-        _ => unreachable!("masked out the upper two bits; qed"),
-    };
-    v.extend(key.as_ref());
-    let r = ss58hash(&v);
-    v.extend(&r[0..2]);
-    bs58::encode(v).into_string()
 }
 
 /// Entry function called from the CLI.
