+++ subp2p-explorer/Cargo.toml
@@ -32,3 +32,6 @@ bytes = { workspace = true }
 codec = { package = "parity-scale-codec", workspace = true, features = ["derive"] }
 primitive-types = { workspace = true, default-features = false, features = ["codec", "scale-info", "serde"] }
 hex = { workspace = true }
+schnorrkel = "0.11.4"
+blake2 = "0.10.4"
+bs58 = { version = "0.5.0", features = ["alloc"] }
