+++ subp2p-explorer/src/util/crypto.rs
@@ -1,0 +1,28 @@
+// Copyright 2023 Alexandru Vasile
+// This file is dual-licensed as Apache-2.0 or GPL-3.0.
+// see LICENSE for license details.
+
+/// Polkadot crypto module.
+pub mod sr25519 {
+    /// Public key for sr25519 keypair implementation.
+    pub type PublicKey = [u8; 32];
+
+    /// Signature generated by signing.
+    pub type Signature = [u8; 64];
+
+    /// Verify that the signature of the message was generated with the public key.
+    pub fn verify<M: AsRef<[u8]>>(sig: &Signature, message: M, pubkey: &PublicKey) -> bool {
+        const SIGNING_CTX: &[u8] = b"substrate";
+
+        let Ok(signature) = schnorrkel::Signature::from_bytes(sig) else {
+            return false;
+        };
+        let Ok(public) = schnorrkel::PublicKey::from_bytes(pubkey) else {
+            return false;
+        };
+
+        public
+            .verify_simple(SIGNING_CTX, message.as_ref(), &signature)
+            .is_ok()
+    }
+}
